organizando menu e main denovo
